+++ xarray/core/variable.py
@@ -1812,7 +1812,7 @@ class Variable(
         return Variable(self.dims, ranked)
 
     def rolling_window(
-        self, dim, window, window_dim, center=False, fill_value=dtypes.NA
+        self, dim, window, window_dim, center=False, fill_value=dtypes.NA, mode=None
     ):
         """
         Make a rolling_window along dim and add a new_dim to the last place.
@@ -1830,6 +1830,9 @@ class Variable(
             of the axis.
         fill_value:
             value to be filled.
+        mode: optional. Default None
+            One of None | 'edge' | 'reflect' | 'symmetric' | 'wrap'
+            If it is not None, fill_value is ignored.
 
         Returns
         -------
@@ -1867,6 +1870,7 @@ class Variable(
                 window=window,
                 center=center,
                 fill_value=fill_value,
+                mode=mode,
             ),
         )