+++ xarray/tests/test_nputils.py
@@ -33,17 +33,35 @@ def test_vindex():
 def test_rolling():
     x = np.array([1, 2, 3, 4], dtype=float)
 
-    actual = rolling_window(x, axis=-1, window=3, center=True, fill_value=np.nan)
+    actual = rolling_window(
+        x, axis=-1, window=3, center=True, fill_value=np.nan, mode=None
+    )
     expected = np.array(
         [[np.nan, 1, 2], [1, 2, 3], [2, 3, 4], [3, 4, np.nan]], dtype=float
     )
     assert_array_equal(actual, expected)
 
-    actual = rolling_window(x, axis=-1, window=3, center=False, fill_value=0.0)
+    actual = rolling_window(
+        x, axis=-1, window=3, center=False, fill_value=0.0, mode=None
+    )
     expected = np.array([[0, 0, 1], [0, 1, 2], [1, 2, 3], [2, 3, 4]], dtype=float)
     assert_array_equal(actual, expected)
 
+    actual = rolling_window(
+        x, axis=-1, window=3, center=True, fill_value=None, mode="wrap"
+    )
+    expected = np.array([[4, 1, 2], [1, 2, 3], [2, 3, 4], [3, 4, 1]], dtype=float)
+    assert_array_equal(actual, expected)
+
+    actual = rolling_window(
+        x, axis=-1, window=3, center=True, fill_value=None, mode="reflect"
+    )
+    expected = np.array([[2, 1, 2], [1, 2, 3], [2, 3, 4], [3, 4, 3]], dtype=float)
+    assert_array_equal(actual, expected)
+
     x = np.stack([x, x * 1.1])
-    actual = rolling_window(x, axis=-1, window=3, center=False, fill_value=0.0)
+    actual = rolling_window(
+        x, axis=-1, window=3, center=True, fill_value=None, mode="reflect"
+    )
     expected = np.stack([expected, expected * 1.1], axis=0)
     assert_array_equal(actual, expected)